Add Logistic Regression algorithm
First commit. For the moment the algo is implemented as multuple
local/global with five steps in total. In the future this will change to
an iterative algorithm with termination condition.
Also, for the moment only the coefficients are returned. In the future a
complete statistics summary will be returned.