Change how masks are computed
Until now masks where computed independently for each pair of variables.
This now changes to a mask that is computed once globally. If ANY variable
has a NaN in some row, this row is discarded for every variable. This is
statistically more sound since in the case of multiple pairs we are
obtaining results having the same number of datapoints.

Additionally, some minor changes in the unit tests:
p-values are now truncated below 0.001 and tests using minimentalstate
are discarded as this is considered a categorical variable.